Move AST values into own compile units
