+++ Makefile.conf
@@ -7,6 +7,7 @@
 
 SOURCES = \
 	ast.cpp \
+	ast_values.cpp \
 	ast_sel_cmp.cpp \
 	ast_sel_unify.cpp \
 	ast_selectors.cpp \
