+++ win/libsass.targets
@@ -11,6 +11,7 @@
   </ItemGroup>
   <ItemGroup Label="LibSass Headers">
     <ClInclude Include="$(LIBSASS_HEADERS_DIR)\ast.hpp" />
+    <ClInclude Include="$(LIBSASS_HEADERS_DIR)\ast_values.hpp" />
     <ClInclude Include="$(LIBSASS_HEADERS_DIR)\ast_selectors.hpp" />
     <ClInclude Include="$(LIBSASS_HEADERS_DIR)\ast_def_macros.hpp" />
     <ClInclude Include="$(LIBSASS_HEADERS_DIR)\memory\SharedPtr.hpp" />
@@ -78,6 +79,7 @@
   </ItemGroup>
   <ItemGroup Label="LibSass Sources">
     <ClCompile Include="$(LIBSASS_SRC_DIR)\ast.cpp" />
+    <ClCompile Include="$(LIBSASS_SRC_DIR)\ast_values.cpp" />
     <ClCompile Include="$(LIBSASS_SRC_DIR)\ast_sel_cmp.cpp" />
     <ClCompile Include="$(LIBSASS_SRC_DIR)\ast_sel_unify.cpp" />
     <ClCompile Include="$(LIBSASS_SRC_DIR)\ast_selectors.cpp" />
